json cmd refactor + wifi_conn cmd
We have refactored the expected command format from the remote user such that we expect to receive a JSON formatted string.

The wifi_conn functionality has also been added. This function allows a user to enter the AP SSID and PWD as part of the "wifi_conn" cmd string and check if the ESP32 can connect and be allocated an IP address.