ping function
This commit covers the addition of the "ping" function, which allows the user to perform a ping if its been previously connected to a Wi-Fi AP (saved credentials).

It has been tested through a variety of combinations of existing commands and cases, but testing has not been exhaustive and would benefit from additional work in the future.