gps_function
We have managed to implement the GPS function.

This requires (scan.c) to run a thread on startup that is attached to CPU0. This thread queries the NEO-6M GPS Module using UART and stores the latest value in a static char array.

The value of the array can then be sent via BLE GATT similarly to how the previous functions have done it.

Note: This commit has significantly affected the (menuconfig) file to allow the code changes to work. Part of these changes are documented in "core_allocation_README.md".